Add temp_env_var context manager
